scripts/requirements.txt
# Appwrite Storage Cleaner Requirements
appwrite>=7.0.0
python-dotenv>=1.0.0
